[Block Library - Query Loop]: Move sticky control to separate file
